requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0

# Technical Analysis
pandas-ta==0.3.14b0
talib-binary==0.4.26

# Oanda API
oandapyV20==0.7.2

# Telegram Bot
python-telegram-bot==20.7

# Task Scheduling
APScheduler==3.10.4

# Web Interface
jinja2==3.1.2
plotly==5.17.0
streamlit==1.28.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Database (optional)
sqlalchemy==2.0.23
aiosqlite==0.19.0
